CC/requirements.txt
Flask==3.0.0
numpy==1.24.2
Pillow==10.1.0
tensorflow==2.13.1
gunicorn==21.2.0
tensorflow_hub==0.15.0